src/lib/performance/ai/AIPerformanceOptimizer.ts
import { PageSpeedMetrics, PerformanceIssue } from '../types';
import { CSSAnalyzer } from '../../styles/monitor/CSSAnalyzer';

export class AIPerformanceOptimizer {
  private static readonly LEARNING_THRESHOLD = 0.8;
  private static readonly OPTIMIZATION_PATTERNS = new Map([
    ['image-heavy', this.optimizeImages],
    ['render-blocking', this.optimizeRendering],
    ['unused-css', this.optimizeCSS],
    ['heavy-javascript', this.optimizeJavaScript]
  ]);

  static analyzePerformancePattern(metrics: PageSpeedMetrics): string[] {
    const patterns: string[] = [];
    
    if (metrics.metrics.LCP > 2500) {
      patterns.push('image-heavy');
    }
    if (metrics.metrics.FCP > 2000) {
      patterns.push('render-blocking');
    }
    if (metrics.metrics.TTI > 3500) {
      patterns.push('heavy-javascript');
    }

    return patterns;
  }

  static async generateOptimizations(metrics: PageSpeedMetrics): Promise<OptimizationSuggestion[]> {
    const patterns = this.analyzePerformancePattern(metrics);
    const suggestions: OptimizationSuggestion[] = [];

    for (const pattern of patterns) {
      const optimizer = this.OPTIMIZATION_PATTERNS.get(pattern);
      if (optimizer) {
        const optimizations = await optimizer(metrics);
        suggestions.push(...optimizations);
      }
    }

    return this.prioritizeSuggestions(suggestions);
  }

  private static async optimizeImages(metrics: PageSpeedMetrics): Promise<OptimizationSuggestion[]> {
    return [
      {
        type: 'image',
        priority: 'high',
        impact: metrics.metrics.LCP > 3000 ? 'high' : 'medium',
        suggestion: 'Utiliser des images WebP avec compression progressive',
        implementation: {
          code: `
            // Exemple de configuration pour Vite
            export default defineConfig({
              build: {
                rollupOptions: {
                  output: {
                    assetFileNames: 'assets/[name].[hash][extname]',
                  },
                },
              },
              plugins: [
                imageOptimizer({
                  quality: 80,
                  formats: ['webp'],
                  progressive: true
                })
              ]
            });
          `,
          file: 'vite.config.ts'
        }
      }
    ];
  }

  private static async optimizeRendering(metrics: PageSpeedMetrics): Promise<OptimizationSuggestion[]> {
    return [
      {
        type: 'code',
        priority: 'high',
        impact: 'high',
        suggestion: 'Implémenter le code splitting pour les composants lourds',
        implementation: {
          code: `
            // Exemple d'implémentation
            const HeavyComponent = lazy(() => import('./HeavyComponent'));
            
            function App() {
              return (
                <Suspense fallback={<Loading />}>
                  <HeavyComponent />
                </Suspense>
              );
            }
          `,
          file: 'src/App.tsx'
        }
      }
    ];
  }

  private static async optimizeCSS(metrics: PageSpeedMetrics): Promise<OptimizationSuggestion[]> {
    const unusedCSS = await CSSAnalyzer.findUnusedStyles([]);
    return [
      {
        type: 'css',
        priority: 'medium',
        impact: unusedCSS.length > 10 ? 'high' : 'medium',
        suggestion: 'Supprimer les styles CSS inutilisés',
        implementation: {
          selectors: unusedCSS
        }
      }
    ];
  }

  private static async optimizeJavaScript(metrics: PageSpeedMetrics): Promise<OptimizationSuggestion[]> {
    return [
      {
        type: 'javascript',
        priority: 'high',
        impact: 'high',
        suggestion: 'Optimiser le bundle JavaScript',
        implementation: {
          code: `
            // Configuration de splitting pour Vite
            export default defineConfig({
              build: {
                rollupOptions: {
                  output: {
                    manualChunks: {
                      vendor: ['react', 'react-dom'],
                      utils: ['./src/utils'],
                      components: ['./src/components']
                    }
                  }
                }
              }
            });
          `,
          file: 'vite.config.ts'
        }
      }
    ];
  }

  private static prioritizeSuggestions(suggestions: OptimizationSuggestion[]): OptimizationSuggestion[] {
    return suggestions.sort((a, b) => {
      const priorityScore = { high: 3, medium: 2, low: 1 };
      const impactScore = { high: 3, medium: 2, low: 1 };
      
      const scoreA = priorityScore[a.priority] * impactScore[a.impact];
      const scoreB = priorityScore[b.priority] * impactScore[b.impact];
      
      return scoreB - scoreA;
    });
  }
}

interface OptimizationSuggestion {
  type: 'image' | 'code' | 'css' | 'javascript';
  priority: 'high' | 'medium' | 'low';
  impact: 'high' | 'medium' | 'low';
  suggestion: string;
  implementation: {
    code?: string;
    file?: string;
    selectors?: string[];
  };
}